sveltekit: Group routes that require a valid revision
