+++ client/web-sveltekit/src/lib/shared.ts
@@ -7,6 +7,8 @@ export {
     isCloneInProgressErrorLike,
     isRepoSeeOtherErrorLike,
     isRepoNotFoundErrorLike,
+    isRevisionNotFoundErrorLike,
+    CloneInProgressError,
 } from '@sourcegraph/shared/src/backend/errors'
 export { SectionID as SearchSidebarSectionID } from '@sourcegraph/shared/src/settings/temporary/searchSidebar'
 export { TemporarySettingsStorage } from '@sourcegraph/shared/src/settings/temporary/TemporarySettingsStorage'
