+++ client/web-sveltekit/src/params/reporev.ts
@@ -1,0 +1,7 @@
+import type { ParamMatcher } from '@sveltejs/kit'
+
+// This ensures that we never consider paths containing /-/ and pointing
+// to non-existing pages as repo name
+export const match = (param => {
+    return !param.includes('/-/')
+}) satisfies ParamMatcher
