+++ client/web-sveltekit/src/routes/+error.svelte
@@ -1,21 +1,24 @@
 <script lang="ts">
+    import { mdiAlertCircle } from '@mdi/js'
+
     import { page } from '$app/stores'
+    import HeroPage from '$lib/HeroPage.svelte'
+    import { isCloneInProgressErrorLike, isRepoNotFoundErrorLike, isRevisionNotFoundErrorLike } from '$lib/shared'
+
+    import CloneInProgressError from './[...repo=reporev]/CloneInProgressError.svelte'
+    import RepoNotFoundError from './[...repo=reporev]/RepoNotFoundError.svelte'
+    import RevisionNotFoundError from './[...repo=reporev]/RevisionNotFoundError.svelte'
 </script>
 
-<div>
-    {#if $page.error?.message === 'enterprise feature'}
-        <h1>Enterprise feature</h1>
-        <p>This is an enterprise-only feature. See ... for more information.</p>
-    {:else}
-        <h1>{$page.status}</h1>
-        <p>
-            {$page.error?.message}
-        </p>
-    {/if}
-</div>
-
-<style lang="scss">
-    div {
-        margin: 1rem;
-    }
-</style>
+{#if isRepoNotFoundErrorLike($page.error)}
+    <RepoNotFoundError repoName={$page.params.repo} viewerCanAdminister={$page.data.user?.siteAdmin ?? false} />
+{:else if isCloneInProgressErrorLike($page.error)}
+    <CloneInProgressError repoName={$page.params.repo} error={$page.error} />
+{:else if isRevisionNotFoundErrorLike($page.error)}
+    <RevisionNotFoundError />
+{:else}
+    <HeroPage title="Error" svgIconPath={mdiAlertCircle}>
+        <!-- TODO: format error message with markdown -->
+        {$page.error?.message ?? 'Unknown Error'}
+    </HeroPage>
+{/if}
