+++ client/web-sveltekit/src/routes/[...repo=reporev]/(validrev)/-/branches/all/+page.ts
@@ -1,0 +1,17 @@
+import { GitRefType } from '$lib/graphql-operations'
+import { queryGitReferences } from '$lib/loader/repo'
+
+import type { PageLoad } from './$types'
+
+export const load: PageLoad = async ({ parent }) => {
+    const { resolvedRevision } = await parent()
+    return {
+        deferred: {
+            branches: queryGitReferences({
+                repo: resolvedRevision.repo.id,
+                type: GitRefType.GIT_BRANCH,
+                first: 20,
+            }).toPromise(),
+        },
+    }
+}
