+++ client/web-sveltekit/src/routes/[...repo=reporev]/(validrev)/-/commit/[...revspec]/DiffSquares.svelte
