+++ client/web-sveltekit/src/routes/[...repo=reporev]/CloneInProgressError.svelte
@@ -1,0 +1,29 @@
+<script lang="ts">
+    import { mdiSourceRepository } from '@mdi/js'
+    import { onMount } from 'svelte'
+
+    import { invalidate } from '$app/navigation'
+    import HeroPage from '$lib/HeroPage.svelte'
+    import { displayRepoName, type CloneInProgressError } from '$lib/shared'
+
+    // TODO: Find a way to make this type stricter
+    export let error: App.Error | null
+    export let repoName: string
+
+    $: progress = (error as CloneInProgressError).progress
+
+    onMount(() => {
+        // Invalidate the repo root every second to trigger an update
+        // and get new progress information.
+        const id = setInterval(() => invalidate('repo:root'), 1000)
+        return () => clearInterval(id)
+    })
+</script>
+
+<HeroPage title={displayRepoName(repoName)} svgIconPath={mdiSourceRepository}>
+    <code><pre>{progress}</pre></code>
+    <!--TODO add DirectImportRepoAlert -->
+</HeroPage>
+
+<style lang="scss">
+</style>
