+++ client/web-sveltekit/src/routes/[...repo=reporev]/RevisionNotFoundError.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+    import HeroPage from '$lib/HeroPage.svelte'
+</script>
+
+<!--TODO: Should this really be rendered as an empty page?  -->
+<HeroPage title="Empty repository">
+    <svg slot="icon" class="mdi-icon" viewBox="0 0 64 64">
+        <title>Icons 400</title>
+        <g>
+            <path
+                d="M50,16h-1.5c-0.3,0-0.5,0.2-0.5,0.5v35c0,0.3-0.2,0.5-0.5,0.5h-27c-0.5,0-1-0.2-1.4-0.6l-0.6-0.6c-0.1-0.1-0.1-0.2-0.1-0.4
+		c0-0.3,0.2-0.5,0.5-0.5H44c1.1,0,2-0.9,2-2V12c0-1.1-0.9-2-2-2H14c-1.1,0-2,0.9-2,2v36.3c0,1.1,0.4,2.1,1.2,2.8l3.1,3.1
+		c1.1,1.1,2.7,1.8,4.2,1.8H50c1.1,0,2-0.9,2-2V18C52,16.9,51.1,16,50,16z M29,44c-1.7,0-3-1.3-3-3c0-1.7,1.3-3,3-3v0
+		c1.7,0,3,1.3,3,3S30.7,44,29,44z M29,16c4.4,0,8,3.1,8,7c0,6.4-5.8,6.2-5.8,11l0,0c0,0.6-0.4,1-1,1H28c-0.6,0-1-0.4-1-1
+		c0-6.6,5.6-6,5.6-11c0-1.7-1.9-3-3.8-3c-1.9,0-3.8,1.3-3.8,3c0,0.6,0.1,1.1,0.3,1.7c0.2,0.5-0.1,1.1-0.6,1.3
+		c-0.1,0-0.2,0.1-0.3,0.1h-2c-0.5,0-0.8-0.3-1-0.7C21.1,24.5,21,23.8,21,23C21,19.1,24.6,16,29,16z"
+            />
+        </g>
+    </svg>
+</HeroPage>
