--- client/web-sveltekit/src/routes/[...repo]/-/commits/+page.ts
@@ -1,11 +1,0 @@
-import { fetchCommits } from '$lib/repo/api/commits'
-
-import type { PageLoad } from './$types'
-
-export const load: PageLoad = ({ parent }) => ({
-    deferred: {
-        commits: parent()
-            .then(({ resolvedRevision }) => fetchCommits(resolvedRevision))
-            .then(result => result?.nodes ?? []),
-    },
-})
