--- client/web-sveltekit/src/routes/[...repo]/-/tags/+page.ts
@@ -1,19 +1,0 @@
-import { isErrorLike } from '$lib/common'
-import { GitRefType } from '$lib/graphql-operations'
-import { queryGitReferences } from '$lib/loader/repo'
-
-import type { PageLoad } from './$types'
-
-export const load: PageLoad = ({ parent }) => ({
-    deferred: {
-        tags: parent().then(({ resolvedRevision }) =>
-            isErrorLike(resolvedRevision)
-                ? null
-                : queryGitReferences({
-                      repo: resolvedRevision.repo.id,
-                      type: GitRefType.GIT_TAG,
-                      first: 20,
-                  }).toPromise()
-        ),
-    },
-})
